Исправления ошибок

+++ apps/training/src/app/training/training.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from "@nestjs/common";
+import { Injectable, NotFoundException } from "@nestjs/common";
 import { TrainingRepository } from "./training.repository";
 import { CreateTrainingDto } from "./dto/create-training.dto";
 import { Training } from "@fit-friends/shared/app-types";
@@ -6,7 +6,7 @@ import { TrainingEntity } from "./training.entity";
 import { TrainingQuery } from "./query/training.query";
 import { UpdateTrainingDto } from "./dto/update-training.dto";
 import { getRandomItem } from "@fit-friends/util/util-core";
-import { DEFAULT_STATIC_IMAGES } from "./training.constant";
+import { DEFAULT_STATIC_IMAGES, TrainingError } from "./training.constant";
 
 @Injectable()
 export class TrainingService {
@@ -21,6 +21,13 @@ export class TrainingService {
   }
 
   async deleteTraining(id: number): Promise<void> {
+
+    const existTraining = await this.getTraining(id)
+
+    if (!existTraining) {
+      throw new NotFoundException(TrainingError.NotExistTraining);
+    }
+
     this.trainingRepository.destroy(id);
   }
 
@@ -37,6 +44,11 @@ export class TrainingService {
   }
 
   async updateTraining(id: number, dto: UpdateTrainingDto): Promise<Training> {
+    const existTraining = await this.getTraining(id)
+
+    if (!existTraining) {
+      throw new NotFoundException(TrainingError.NotExistTraining);
+    }
     return this.trainingRepository.update(id, new TrainingEntity({...dto }))
   }
